Added heads for all the pages and extracted list logic into list items
