+++ seasonal/pages/index.js
@@ -3,6 +3,7 @@ import styles from "../styles/Home.module.css";
 import { NavBar } from "../components/NavBar.js";
 import { SearchBar } from "../components/SearchBar.js";
 import { RandomFiveArray } from '../components/RandomFiveArray.js'
+
 //import Button from '../components/Button';
