+++ seasonal/pages/results.js
@@ -1,36 +1,29 @@
 import React from "react";
-import { useRouter } from "next/router";
-import { useState, useEffect } from "react";
-import Button from "../components/Button.js";
+import Head from "next/head";
+import { NavBar } from "../components/NavBar.js";
+import { SearchBar } from "../components/SearchBar.js";
+import { ResultsItem } from "../components/ResultsItem.js"; 
+import styles from "../styles/Card.module.css";
 
 const searchString = "https://seasonality-server-new.herokuapp.com/produce";
 
 export default function Results() {
-  const router = useRouter();
-  const [searchResults, setSearchResults] = useState({ payload: [] });
 
-  useEffect(() => {
-    const fetchData = async () => {
-      console.log(JSON.stringify(router.query));
-      let modifiedMonth = router.query.month;
-      const data = await fetch(`${searchString}?month=${modifiedMonth}`);
-      let result = await data.json();
-      setSearchResults(result);
-      console.log(
-        `this is the search results console log: ${JSON.stringify(result)}`
-      );
-    };
-    if (!router.isReady) return;
-    fetchData()
-      // make sure to catch any error
-      .catch(console.error);
-  }, [router.isReady]);
+return(
+<div>
+  <Head>
+        <title>Seasonal</title>
+        <meta name="description" content="Search by months results page" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+  
+  <NavBar/>
+ <ResultsItem/> 
+ <footer className={styles.footer}>
+        <SearchBar />
+      </footer>
+ </div>
+)
 
-  if (searchResults.payload.length > 0) {
-    return searchResults.payload.map((result) => {
-      return <Button text={result.name} />;
-    });
-  } else {
-    return <p>Loading</p>;
-  }
+
 }
